Fix HTML5 paste event handling
MS Edge has a bug with clipboardData.types not being an Array and thus not having indexOf, so the handler fails with an exception. But we can just `getData` and check if it's an empty string.

Also let's do `preventDefault()` before doing anything else, because both Clipboard.text setter and onTextInput handlers can throw exceptions, but we still want to prevent the default handling.